Upgrade cpp docs Sphinx/breathe/exhale to latest version
I was hoping that exhale had gotten incremental recompilation
in its latest version, but experimentally this does not seem
to have been the case.  Still, I had gotten the whole shebang
to be working on the latest version of these packages, so might
as well land the upgrade.  There was one bug in Optional.h that
I had to fix; see the cited bug report.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 14b087ad2871aac77e42e8e86ea68b26237807a3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41312